Update leaflet.js
Have added a leaflet_OnGoogleMapsLoaded_defer wrapper around the google.load() execution - since we don't know when jsapi will be available for usage. We also don't know when jQuery will be avaiable, so we have a sanity check for that as well.
It's the same code, only it checks for jquery (it's required), and uses a document.write hook to perform a jquery.fn.append instead. It restores it directly afterward so as to not affect anyone else's code.
You'll notice at the end of it's execution, it calls leaflet_MapsReadyCallback();  
That's the function that performs the actual leaflet map initialization, which too has been wrapped.